+++ tests/test_models.py
@@ -29,6 +29,7 @@ def test_movieinfo_from_movie_search():
     data = {
         "titleId": "tt0133093",
         "titleText": "The Matrix",
+        "originalTitleText" : "The Matrix",
         "primaryImage": {"url": "http://example.com/matrix.jpg"},
         "releaseYear": "1999",
         "titleType": {"id":"movie"},